Resolve launcher PR preflight issues
