+++ crates/oasis7_client_launcher/src/main_tests.rs
@@ -57,11 +57,6 @@ fn hosted_strong_auth_env_lock() -> &'static Mutex<()> {
     LOCK.get_or_init(|| Mutex::new(()))
 }
 
-fn launcher_screenshot_modal_env_lock() -> &'static Mutex<()> {
-    static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-    LOCK.get_or_init(|| Mutex::new(()))
-}
-
 fn clear_hosted_strong_auth_env() {
     for name in [
         "OASIS7_HOSTED_STRONG_AUTH_PUBLIC_KEY",
@@ -1147,61 +1142,6 @@ fn collect_required_config_issues_accepts_bundle_relative_web_path_from_launcher
     assert!(!issues.contains(&ConfigIssue::LauncherBinMissing));
 
     let _ = fs::remove_dir_all(bundle_root);
-}
-
-#[test]
-fn heavy_modal_windows_render_one_real_egui_frame() {
-    let ctx = egui::Context::default();
-    let mut app = ClientLauncherApp::default();
-    app.config.chain_enabled = true;
-    app.chain_runtime_status = ChainRuntimeStatus::Ready;
-    app.transfer_window_open = true;
-    app.explorer_window_open = true;
-    app.peer_details_window_open = true;
-
-    let _ = ctx.run(egui::RawInput::default(), |ctx| {
-        app.show_transfer_window(ctx);
-        app.show_explorer_window(ctx);
-        app.show_peer_details_window(ctx);
-    });
-
-    assert!(app.transfer_window_open);
-    assert!(app.explorer_window_open);
-    assert!(app.peer_details_window_open);
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-#[test]
-fn screenshot_modal_override_seeds_without_opening_real_windows() {
-    let _guard = launcher_screenshot_modal_env_lock()
-        .lock()
-        .expect("screenshot modal env lock");
-    let previous = std::env::var(OASIS7_CLIENT_LAUNCHER_SCREENSHOT_MODAL_ENV).ok();
-    std::env::set_var(OASIS7_CLIENT_LAUNCHER_SCREENSHOT_MODAL_ENV, "transfer");
-
-    let mut app = ClientLauncherApp::default();
-    app.onboarding_state.open = true;
-
-    app.apply_screenshot_modal_override();
-
-    if let Some(previous) = previous {
-        std::env::set_var(OASIS7_CLIENT_LAUNCHER_SCREENSHOT_MODAL_ENV, previous);
-    } else {
-        std::env::remove_var(OASIS7_CLIENT_LAUNCHER_SCREENSHOT_MODAL_ENV);
-    }
-
-    assert!(!app.onboarding_state.open);
-    assert!(!app.config_window_open);
-    assert!(!app.feedback_window_open);
-    assert!(!app.transfer_window_open);
-    assert!(!app.explorer_window_open);
-    assert!(!app.peer_details_window_open);
-    assert!(app.config.chain_enabled);
-    assert!(matches!(
-        app.chain_runtime_status,
-        ChainRuntimeStatus::Ready
-    ));
-    assert_eq!(app.transfer_draft.amount, "12");
 }
 
 fn make_temp_dir(label: &str) -> PathBuf {
